# Conflicts:
#	gieBackend/output/classes/META-INF/persistence.xml
#	gieBackend/output/classes/org/gieback/App.class
#	gieBackend/output/classes/org/gieback/Controller/AchatController.class
#	gieBackend/output/classes/org/gieback/Controller/CommandeController.class
#	gieBackend/output/classes/org/gieback/Controller/EntrepriseController.class
#	gieBackend/output/classes/org/gieback/Controller/PersonneController.class
#	gieBackend/output/classes/org/gieback/Controller/ProductController.class
#	gieBackend/output/classes/org/gieback/DAO/AchatDao.class
#	gieBackend/output/classes/org/gieback/DAO/CommandeDao.class
#	gieBackend/output/classes/org/gieback/DAO/EntrepriseDao.class
#	gieBackend/output/classes/org/gieback/DAO/IAchatDao.class
#	gieBackend/output/classes/org/gieback/DAO/IEntrepriseDao.class
#	gieBackend/output/classes/org/gieback/DAO/IPersonneDao.class
#	gieBackend/output/classes/org/gieback/DAO/IProductDao.class
#	gieBackend/output/classes/org/gieback/DAO/PersonneDao.class
#	gieBackend/output/classes/org/gieback/DAO/ProductDao.class
#	gieBackend/output/classes/org/gieback/Entity/Achat.class
#	gieBackend/output/classes/org/gieback/Entity/AchatDetail.class
#	gieBackend/output/classes/org/gieback/Entity/Category.class
#	gieBackend/output/classes/org/gieback/Entity/Commande.class
#	gieBackend/output/classes/org/gieback/Entity/Contact.class
#	gieBackend/output/classes/org/gieback/Entity/EtatCommande.class
#	gieBackend/output/classes/org/gieback/Entity/Product.class
#	gieBackend/output/classes/org/gieback/Service/AchatService.class
#	gieBackend/output/classes/org/gieback/Service/CommandeService.class
#	gieBackend/output/classes/org/gieback/Service/EntrepriseService.class
#	gieBackend/output/classes/org/gieback/Service/IAchatService.class
#	gieBackend/output/classes/org/gieback/Service/IEntrepriseService.class
#	gieBackend/output/classes/org/gieback/Service/IPersonneService.class
#	gieBackend/output/classes/org/gieback/Service/IProductService.class
#	gieBackend/output/classes/org/gieback/Service/PersonneService.class
#	gieBackend/output/classes/org/gieback/Service/ProductService.class
#	gieBackend/src/main/java/org/gieback/Controller/CommandeController.java
#	gieBackend/src/main/java/org/gieback/DAO/CommandeDao.java
#	gieBackend/src/main/java/org/gieback/Entity/Achat.java
#	gieBackend/src/main/java/org/gieback/Entity/Category.java
#	gieBackend/src/main/java/org/gieback/Service/CommandeService.java
#	gieBackend/src/main/java/org/gieback/Service/ICommandeService.java
#	gieBackend/src/main/resources/META-INF/persistence.xml
#	gieFrontend/src/main/java/module-info.java
#	gieFrontend/src/main/java/org/giefront/AddCommandeController.java
#	gieFrontend/src/main/java/org/giefront/DTO/Achat.java
#	gieFrontend/src/main/java/org/giefront/DTO/AchatDetail.java
#	gieFrontend/src/main/java/org/giefront/DTO/Category.java
#	gieFrontend/src/main/java/org/giefront/DTO/Commande.java
#	gieFrontend/src/main/java/org/giefront/DTO/EtatCommande.java
#	gieFrontend/src/main/java/org/giefront/DTO/Product.java
#	gieFrontend/src/main/java/org/giefront/EditDeleteCell.java
#	gieFrontend/src/main/java/org/giefront/EditDeleteCellEnt.java
#	gieFrontend/src/main/java/org/giefront/EntrepriseController.java
#	gieFrontend/src/main/java/org/giefront/EntrepriseModifController.java
#	gieFrontend/src/main/java/org/giefront/FournisseurPersoController.java
#	gieFrontend/src/main/java/org/giefront/FournisseurPersoControlleur.java
#	gieFrontend/src/main/java/org/giefront/ProductController.java
#	gieFrontend/src/main/java/org/giefront/Service/AchatService.java
#	gieFrontend/src/main/java/org/giefront/Service/CommandeService.java
#	gieFrontend/src/main/java/org/giefront/Service/EntrepriseService.java
#	gieFrontend/src/main/java/org/giefront/Service/ProductService.java
#	gieFrontend/src/main/java/org/giefront/TestFront.java
#	gieFrontend/src/main/java/org/giefront/controller/CommandeController.java
#	gieFrontend/src/main/resources/org/Interfaces/addCommande.fxml
#	gieFrontend/src/main/resources/org/Interfaces/commandUI.fxml
#	gieFrontend/src/main/resources/org/Interfaces/fournisseurEntro.fxml
#	gieFrontend/src/main/resources/org/Interfaces/fournisseurPerso.fxml

+++ gieFrontend/src/main/java/org/giefront/AdvancedSearchE.java
@@ -116,7 +116,7 @@ public class AdvancedSearchE implements Initializable {
     public void onDelete(ActionEvent event) throws IOException {
         Entreprise selectedEntreprise = tableView_E.getSelectionModel().getSelectedItem();
         if (selectedEntreprise != null) {
-            e.deleteEntreprise(Math.toIntExact(selectedEntreprise.getId()));
+            e.deleteEntreprise((long) Math.toIntExact(selectedEntreprise.getId()));
             tableView_E.getItems().remove(selectedEntreprise);
             showAlert("Suppression reussie");
         }
